hotelbackend/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-cors-headers==3.4.0
Pillow==7.2.0